Made admin panel work with auth=none

+++ php/class.RestUsersUserDriver.php
@@ -24,10 +24,13 @@ class RestUsersUserDriver extends RestDriver
 		}
 	function head()
 		{
-		$this->core->db->selectDb($this->core->database->database);
-		$this->core->db->query('SELECT users.id as userid, login, firstname, lastname, email, organization, groups.name as groupname, groups.id as groupid, lastconnection FROM users LEFT JOIN groups ON groups.id=users.group WHERE login="'.$this->request->uriNodes[1].'"');
-		if(!$this->core->db->numRows())
-			throw new RestException(RestCodes::HTTP_410,'This user doesn\'t exist, uh ?');
+		if($this->core->server->auth!='none')
+			{
+			$this->core->db->selectDb($this->core->database->database);
+			$this->core->db->query('SELECT users.id as userid, login, firstname, lastname, email, organization, groups.name as groupname, groups.id as groupid, lastconnection FROM users LEFT JOIN groups ON groups.id=users.group WHERE login="'.$this->request->uriNodes[1].'"');
+			if(!$this->core->db->numRows())
+				throw new RestException(RestCodes::HTTP_410,'This user doesn\'t exist, uh ?');
+			}
 		return new RestResponse(
 			RestCodes::HTTP_200,
 			array('Content-Type'=>'text/varstream')
@@ -38,21 +41,36 @@ class RestUsersUserDriver extends RestDriver
 		$response=$this->head();
 		$response->content=new stdClass();
 		$response->content->user=new stdClass();
-		$response->content->user->userId = $this->core->db->result('userid');
-		$response->content->user->login = $this->core->db->result('login');
-		$response->content->user->firstName = $this->core->db->result('firstname');
-		$response->content->user->lastName = $this->core->db->result('lastname');
-		$response->content->user->email = $this->core->db->result('email');
-		$response->content->user->organization = $this->core->db->result('organization');
-		$response->content->user->groupName = $this->core->db->result('groupname');
-		$response->content->user->groupId = $this->core->db->result('groupid');
-		if($this->queryParams->type!='restricted')
-			$response->content->user->lastconnection = $this->core->db->result('lastconnection');
+		if($this->core->server->auth=='none')
+			{
+			$response->content->user->userId = 1;
+			$response->content->user->login = 'webmaster';
+			$response->content->user->firstName = 'Unknow';
+			$response->content->user->lastName = 'Unknow';
+			$response->content->user->organization = 1;
+			$response->content->user->groupName = 'webmasters';
+			$response->content->user->groupId = 1;
+			}
+		else
+			{
+			$response->content->user->userId = $this->core->db->result('userid');
+			$response->content->user->login = $this->core->db->result('login');
+			$response->content->user->firstName = $this->core->db->result('firstname');
+			$response->content->user->lastName = $this->core->db->result('lastname');
+			$response->content->user->email = $this->core->db->result('email');
+			$response->content->user->organization = $this->core->db->result('organization');
+			$response->content->user->groupName = $this->core->db->result('groupname');
+			$response->content->user->groupId = $this->core->db->result('groupid');
+			if($this->queryParams->type!='restricted')
+				$response->content->user->lastconnection = $this->core->db->result('lastconnection');
+			}
 		$response->setHeader('Content-Type','text/varstream');
 		return $response;
 		}
 	function put()
 		{
+		if($this->core->server->auth=='none')
+			throw new RestException(RestCodes::HTTP_400,'Unable to modify the default user');
 		try
 			{
 			$response=$this->head();
@@ -86,6 +104,8 @@ class RestUsersUserDriver extends RestDriver
 		}
 	function delete()
 		{
+		if($this->core->server->auth=='none')
+			throw new RestException(RestCodes::HTTP_400,'Unable to delete the default user');
 		$this->core->db->query('DELETE FROM users WHERE login="'.$this->request->uriNodes[1].'"');
 		return new RestResponse(RestCodes::HTTP_200,array('Content-Type','text/varstream'));
 		}
